requirements.txt
pyTelegramBotAPI 
python-dotenv 
requests
telebot